Moved basic types to a Common module.
The types `Generator`, `Height`, `SliceIndex` didn't belong in the `Rewrite` or `Diagram`
module but are too simple to justify a module of their own, so they have been moved to
a `Common` module. The boundary part of the `SliceIndex` type has also been factored out
so that we're able to talk about source and target boundaries in isolation.