Added 2D layout solver.
This solver does naively generate all average and distance constraints, but uses a
preprocessing phase that identifies solver variables that should always have the
same value. By using the natural ordering of the points which is compatible with
the iteration order, no union find is neccessary.

A test for left and right associated multiplications is provided, but more tests
especially including scalars are needed. Also this has not been tested for performance.

The API exposed by the `Layout` module is preliminary.