Added a `Layout` type to represent a solved layout.
A solved layout was represented directly by a function from points to their
positions and the solver thus was a function from a diagram and a point to a
position. This function was intended to be partially applied to the diagram
first and the result reused for multiple points to avoid solving the layout
again. This wasn't entirely apparent from the API.

Also this hopefully fixed the pre-commit hook.